requirements_api.txt
# Bibliothèques de base pour le traitement des données
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2

# Connectivité à la base de données PostgreSQL
psycopg2==2.9.9
sqlalchemy>=1.4.28,<2.0

# Client API Binance pour récupérer les données des cryptomonnaies
python-binance==1.0.16
requests==2.28.1

# Framework Web & FastAPI pour créer l'API
fastapi==0.103.0  # Compatible avec Pydantic 2.x
uvicorn==0.18.3

# Validation des e-mails (compatible avec FastAPI et API)
email-validator>=2.0

# WebSocket pour le streaming des données Binance
websockets==10.3

# Bibliothèques pour les graphiques (si nécessaires dans l'API)
plotly==5.10.0
dash==2.7.0
matplotlib==3.7.1

